fix(ast): temporary fix tsify not generating some typings
seems like a bug in tsify
