+++ crates/oxc_ast/src/ast/mod.rs
@@ -1,9 +1,8 @@
 //! AST Definitions
 
 mod js;
-mod jsdoc;
 mod jsx;
 mod literal;
 mod ts;
 
-pub use self::{js::*, jsdoc::*, jsx::*, literal::*, ts::*};
+pub use self::{js::*, jsx::*, literal::*, ts::*};
